+++ trunk/RenderSystems/GL/include/OgreGLRenderSystem.h
@@ -200,7 +200,7 @@ namespace Ogre {
         /** See
           RenderSystem
          */
-        void convertColourValue(const ColourValue& colour, unsigned long* pDest);
+        void convertColourValue(const ColourValue& colour, uint32* pDest);
         /** See
           RenderSystem
          */
